# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Casilleros ESCOM",
    "version": "1.0.0",
    "description": "Documentación de la API para la gestión de casilleros en ESCOM."
  },
  "servers": [
    { "url": "http://localhost:3001" }
  ],
  "paths": {
    "/usuarios": {
      "post": {
        "summary": "Registrar un nuevo usuario (alumno o coordinador)",
        "tags": ["Usuarios"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "correo": { "type": "string", "example": "alumno@escom.ipn.mx" },
                  "contrasena": { "type": "string", "example": "123456" },
                  "rol": { "type": "string", "enum": ["alumno", "coordinador"] }
                },
                "required": ["correo", "contrasena", "rol"]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuario registrado correctamente",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_usuario": { "type": "integer" },
                    "correo": { "type": "string" },
                    "rol": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Faltan campos requeridos" },
          "409": { "description": "El correo ya está registrado" },
          "500": { "description": "Error al registrar usuario" }
        }
      }
    },
    "/usuarios/login": {
      "post": {
        "summary": "Iniciar sesión de usuario",
        "tags": ["Usuarios"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "correo": { "type": "string", "example": "alumno@escom.ipn.mx" },
                  "contrasena": { "type": "string", "example": "123456" }
                },
                "required": ["correo", "contrasena"]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Inicio de sesión exitoso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id_usuario": { "type": "integer" },
                    "correo": { "type": "string" },
                    "rol": { "type": "string" }
                  }
                }
              }
            }
          },
          "400": { "description": "Faltan campos requeridos" },
          "401": { "description": "Credenciales inválidas" },
          "500": { "description": "Error al iniciar sesión" }
        }
      }
    },
...
    "/solicitudes/usuario/{id_usuario}": {
      "get": {
        "summary": "Obtener solicitudes de un usuario (alumno)",
        "tags": ["Solicitudes"],
        "parameters": [
          {
            "name": "id_usuario",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de solicitudes del usuario",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Solicitud" }
                }
              }
            }
          },
          "500": { "description": "Error al obtener solicitudes del usuario" }
        }
      }
    },
...
    "/asignaciones/usuario/{id_usuario}": {
      "get": {
        "summary": "Obtener asignaciones de un usuario",
        "tags": ["Asignaciones"],
        "parameters": [
          {
            "name": "id_usuario",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Lista de asignaciones del usuario" },
          "500": { "description": "Error al obtener asignaciones" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [
    { "bearerAuth": [] }
  ]
}